+++ lending/src/sdk/sdk.ts
@@ -66,9 +66,9 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "3.0.0";
-    sdkVersion = "0.5.1";
-    genVersion = "2.231.0";
-    userAgent = "speakeasy-sdk/typescript 0.5.1 2.231.0 3.0.0 lending";
+    sdkVersion = "0.6.0";
+    genVersion = "2.248.6";
+    userAgent = "speakeasy-sdk/typescript 0.6.0 2.248.6 3.0.0 lending";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);