+++ lending/RELEASES.md
@@ -109,3 +109,13 @@ Based on:
 - [typescript v0.5.1] lending
 ### Releases
 - [NPM v0.5.1] https://www.npmjs.com/package/lending/v/0.5.1 - lending
+
+## 2024-02-05 01:28:07
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Lending.yaml
+- Speakeasy CLI 1.166.0 (2.248.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.6.0] lending
+### Releases
+- [NPM v0.6.0] https://www.npmjs.com/package/lending/v/0.6.0 - lending